Large refactor
- Add username and password auth
- Implement MQTT availability topics
- Ensure we clean all device and sensor names that make it to MQTT topic names
- Add more data to MQTT devices vs "unknown"
- Don't recreate a bunch of clients
- Make it one giant file: it's only 300 LoC
- Actually use `INTERVAL` env var to set publish interval
- Don't republish the discovery message every time there's a new sensor value published
